+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Stop doing a get request for every product on properties and specification groups resolvers, use skuSpecifications filed from catalog to get field id.
 
 ## [0.8.4] - 2020-07-09
 ### Fixed